# src/management-system/src/backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PROCEED Management System API",
    "description": "### About us\nPROCEED is a decentralized Business Process Management System (dBPMS) specialized on the creation, execution and monitoring of individual IoT processes. The execution is done decentralized without a controlling instance. Instead there are multiple machines that cooperate with each other to achieve the designed process goal. Thereby, all machines communicate their capabilities to the vicinity and the process is taking the best fitting one for its next execution ##step.\n\n### More resources: \n- [Documentation](https://docs.proceed-labs.org/)\n- [Wiki](https://github.com/PROCEED-Labs/proceed/wiki)",
    "contact": {
      "email": "proceed@snet.tu-berlin.de",
      "url": "mailto:proceed@snet.tu-berlin.de",
      "name": "proceed"
    },
    "license": {
      "name": "MIT License",
      "url": "https://github.com/PROCEED-Labs/proceed/blob/main/LICENSE"
    },
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "https://localhost:33083/api",
      "description": "Managent system in development"
    },
    {
      "url": "https://proceed.snet.tu-berlin.de/api",
      "description": ""
    }
  ],
...
  "security": [
    {
      "id": [],
      "x-csrf-token": [],
      "x-csrf": []
    },
    {}
  ],
  "paths": {
...
    "/users/{id}": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "id",
          "in": "path",
          "required": true,
          "description": "Id of a user"
        }
      ],
      "get": {
        "tags": ["Users"],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/userResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request (User probably doesn't exist)",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401_unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/403_validationFailed"
          }
        },
        "operationId": "getUserById",
        "description": "Get a specific user."
      },
      "put": {
        "tags": ["Users"],
        "operationId": "updateUserById",
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401_unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/403_validationFailed"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/userDataPut"
              }
            }
          }
        },
        "description": "Update a specific user."
      },
      "delete": {
        "tags": ["Users"],
        "operationId": "deleteUserById",
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401_unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/403_validationFailed"
          }
        },
        "description": "Delete a specific user."
      }
    },
...
    "/roles": {
      "get": {
        "tags": ["Roles"],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/roleResponse"
                }
              }
            }
          },
          "204": {
            "description": "No Content",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "minItems": 0,
                  "maxItems": 0,
                  "items": {}
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401_unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/403_validationFailed"
          }
        },
        "operationId": "getRole",
        "description": "Get all roles of the authorization system."
      },
      "post": {
        "tags": ["Roles"],
        "operationId": "postRole",
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/roleResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401_unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/403_validationFailed"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/roleData"
                  }
                ],
                "required": [
                  "name",
                  "description",
                  "note",
                  "permissions",
                  "expiration",
                  "members",
                  "default"
                ]
              }
            }
          }
        },
        "description": "Post a new role."
      }
    },
    "/roles/{id}": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "id",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "tags": ["Roles"],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/roleResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400_error_string_or_message"
          },
          "401": {
            "$ref": "#/components/responses/401_unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/403_validationFailed"
          }
        },
        "operationId": "getRoleById",
        "description": "Get a specific role."
      },
      "put": {
        "tags": ["Roles"],
        "operationId": "updateRoleById",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/roleResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400_error_string_or_message"
          },
          "401": {
            "$ref": "#/components/responses/401_unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/403_validationFailed"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/roleData"
              }
            }
          }
        },
        "description": "Update a specific role."
      },
      "delete": {
        "tags": ["Roles"],
        "operationId": "deleteRoleById",
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "$ref": "#/components/responses/400_Error_Json"
          },
          "401": {
            "$ref": "#/components/responses/401_unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/403_validationFailed"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        },
        "description": "Delete a specific role."
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "userMetaData": {
        "title": "user",
        "type": "object",
        "properties": {
          "created_at": {
            "type": "string"
          },
          "email_verified": {
            "type": "boolean"
          },
          "updated_at": {
            "type": "string"
          },
          "id": {
            "type": "string"
          }
        }
      },
      "userDataPut": {
        "title": "user",
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "username": {
            "type": "string"
          },
          "firstName ": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          }
        },
        "required": ["email", "username", "firstName ", "lastName"]
      },
      "userData": {
        "title": "user",
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string"
          },
          "picture": {
            "type": "string",
            "format": "uri"
          },
          "username": {
            "type": "string"
          },
          "lastName ": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          }
        }
      },
      "userResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/userMetaData"
          },
          {
            "$ref": "#/components/schemas/userData"
          }
        ],
        "required": [
          "created_at",
          "email",
          "email_verified",
          "name",
          "picture",
          "updated_at",
          "username",
          "lastName",
          "firstName",
          "id"
        ]
      },
      "PermissionNumber": {
        "title": "PermissionNumber",
        "type": "number",
        "description": "- example\n- ezample 2"
      },
      "roleData": {
        "title": "role",
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "permissions": {
            "type": "object",
            "properties": {
              "Process": {
                "$ref": "#/components/schemas/PermissionNumber"
              },
              "Project": {
                "$ref": "#/components/schemas/PermissionNumber"
              },
              "Template": {
                "$ref": "#/components/schemas/PermissionNumber"
              },
              "Task": {
                "$ref": "#/components/schemas/PermissionNumber"
              },
              "Machine": {
                "$ref": "#/components/schemas/PermissionNumber"
              },
              "Execution": {
                "$ref": "#/components/schemas/PermissionNumber"
              },
              "Role": {
                "$ref": "#/components/schemas/PermissionNumber"
              },
              "User": {
                "$ref": "#/components/schemas/PermissionNumber"
              },
              "Setting": {
                "$ref": "#/components/schemas/PermissionNumber"
              },
              "EnvConfig": {
                "$ref": "#/components/schemas/PermissionNumber"
              },
              "All": {
                "$ref": "#/components/schemas/PermissionNumber"
              }
            }
          },
          "expiration": {
            "type": "string"
          },
          "members": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "userId": {
                  "type": "string"
                },
                "username": {
                  "type": "string"
                },
                "fisrtName": {
                  "type": "string"
                },
                "lastName": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                }
              },
              "required": ["userId", "username", "fisrtName", "lastName", "email"]
            }
          },
          "default": {
            "type": "boolean"
          }
        }
      },
      "roleMetaData": {
        "title": "role",
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdOn": {
            "type": "string"
          },
          "lastEdited": {
            "type": "string"
          }
        }
      },
      "roleResponse": {
        "title": "roleRespnse",
        "allOf": [
          {
            "$ref": "#/components/schemas/roleData"
          },
          {
            "$ref": "#/components/schemas/roleMetaData"
          }
        ],
        "required": [
          "name",
          "description",
          "note",
          "permissions",
          "expiration",
          "members",
          "id",
          "default",
          "createdOn",
          "lastEdited"
        ]
      },
...
    },
    "securitySchemes": {
      "x-csrf-token": {
        "name": "Cross-site request forgery token",
        "type": "apiKey",
        "in": "header",
        "description": "Cross-site request forgery token obtained after logging in."
      },
      "x-csrf": {
        "name": "Cross-site request forgery enabled",
        "type": "apiKey",
        "in": "header",
        "description": "Needs to be set to '1' in order for the x-csrf-token to be acknowledged by the server and make authenticated requests."
      },
      "id": {
        "name": "Session id",
        "type": "apiKey",
        "in": "cookie",
        "description": "Session id given by the server after logging in."
      }
    },
...
    "responses": {
      "400_Error_Json": {
        "description": "Example response",
        "content": {
          "application/json": {
            "schema": {
              "type": "string",
              "description": "Error description."
            }
          },
          "text/html": {
            "schema": {
              "type": "string"
            }
          }
        },
        "headers": {}
      },
...
      "400_error_string_or_message": {
        "description": "400 Error.",
        "content": {
          "application/json": {
            "schema": {
              "oneOf": [
                {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                {
                  "type": "string"
                }
              ]
            }
          }
        }
      },
...
      "401_unauthenticated": {
        "description": "Authentication error.",
        "content": {
          "text/html": {
            "schema": {
              "type": "string",
              "description": "Authentication error."
            }
          }
        }
      },
      "403_validationFailed": {
        "description": "Request validation failed.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "x-examples": {
                "Example 1": {
                  "aditionalProperties": true
                }
              },
              "description": "Request validation error."
            }
          },
          "text/html": {
            "schema": {
              "type": "string",
              "description": "Authorization error."
            }
          }
        }
      }
    }
  }
}